# Switch from md5 to xxhash for score