Bench:

```
benchmark                   old ns/op      new ns/op      delta
BenchmarkOpenIndex-4        82574461       65227740       -21.01%
BenchmarkRebuildIndex-4     1379846122     1370667311     -0.67%
BenchmarkOpen-4             81791544       65762731       -19.60%
BenchmarkWrite-4            22247          19775          -11.11%
BenchmarkRead-4             1965           1661           -15.47%

benchmark                   old allocs     new allocs     delta
BenchmarkOpenIndex-4        9616           9666           +0.52%
BenchmarkRebuildIndex-4     245445         245524         +0.03%
BenchmarkOpen-4             9616           9693           +0.80%
BenchmarkWrite-4            2              2              +0.00%
BenchmarkRead-4             1              1              +0.00%

benchmark                   old bytes     new bytes     delta
BenchmarkOpenIndex-4        41077400      31396703      -23.57%
BenchmarkRebuildIndex-4     48616296      38929176      -19.93%
BenchmarkOpen-4             41072258      31403386      -23.54%
BenchmarkWrite-4            1292          1266          -2.01%
BenchmarkRead-4             32            32            +0.00%
```